+++ controllers/controller_common.go
@@ -2,9 +2,11 @@ package controllers
 
 import (
 	"fmt"
-	"github.com/konpyutaika/nifikop/api/v1"
 	"strings"
 	"time"
+
+	v1 "github.com/konpyutaika/nifikop/api/v1"
+	"github.com/konpyutaika/nifikop/api/v1alpha1"
 
 	"github.com/go-logr/logr"
 	"github.com/konpyutaika/nifikop/pkg/util/clientconfig"
@@ -164,6 +166,17 @@ func GetUserRefNamespace(ns string, ref v1.UserReference) string {
 		return ns
 	}
 	return userNamespace
+}
+
+// getComponentRefNamespace returns the expected namespace for a Nifi cluster
+// referenced by a user/dataflow CR. It takes the namespace of the CR as the first
+// argument and the reference itself as the second.
+func GetComponentRefNamespace(ns string, ref v1alpha1.ComponentReference) string {
+	componentNamespace := ref.Namespace
+	if componentNamespace == "" {
+		return ns
+	}
+	return componentNamespace
 }
 
 func GetLogConstructor(mgr manager.Manager, obj runtime.Object) (func(*reconcile.Request) logr.Logger, error) {