+++ controllers/nificluster_controller.go
@@ -43,8 +43,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-var clusterFinalizer = "nificlusters.nifi.konpyutaika.com/finalizer"
-var clusterUsersFinalizer = "nificlusters.nifi.konpyutaika.com/users"
+var clusterFinalizer string = fmt.Sprintf("nificlusters.%s/finalizer", v1.GroupVersion.Group)
+var clusterUsersFinalizer string = fmt.Sprintf("nificlusters.%s/users", v1.GroupVersion.Group)
 
 // NifiClusterReconciler reconciles a NifiCluster object
 type NifiClusterReconciler struct {