+++ pkg/errorfactory/errorfactory.go
@@ -77,7 +77,7 @@ type NifiFlowUpdateRequestNotFound struct{ error }
 // NifiFlowControllerServiceScheduling states that the flow's controller service are still scheduling
 type NifiFlowControllerServiceScheduling struct{ error }
 
-// NifiFlowSyncing states that the flow's controller service are still scheduling
+// NifiFlowSyncing states that the flow is still syncing
 type NifiFlowSyncing struct{ error }
 
 // NifiFlowScheduling states that the flow is still scheduling
@@ -88,6 +88,12 @@ type NifiReportingTasksValidating struct{ error }
 
 // NifiReportingTasksInvalid states that the reporting task is invalid
 type NifiReportingTasksInvalid struct{ error }
+
+// NifiConnectionSyncing states that the connection is still syncing
+type NifiConnectionSyncing struct{ error }
+
+// NifiConnectionDeleting states that the connection is still deleting
+type NifiConnectionDeleting struct{ error }
 
 // New creates a new error factory error
 func New(t interface{}, err error, msg string, wrapArgs ...interface{}) error {