+++ pkg/nificlient/inputport.go
@@ -21,3 +21,20 @@ func (n *nifiClient) UpdateInputPortRunStatus(id string, entity nigoapi.PortRunS
 
 	return &processor, nil
 }
+
+func (n *nifiClient) GetInputPort(id string) (*nigoapi.PortEntity, error) {
+	// Get nigoapi client, favoring the one associated to the coordinator node.
+	client, context := n.privilegeCoordinatorClient()
+	if client == nil {
+		n.log.Error("Error during creating node client", zap.Error(ErrNoNodeClientsAvailable))
+		return nil, ErrNoNodeClientsAvailable
+	}
+
+	// Request on Nifi Rest API to update the input port informations
+	port, rsp, body, err := client.InputPortsApi.GetInputPort(context, id)
+	if err := errorUpdateOperation(rsp, body, err, n.log); err != nil {
+		return nil, err
+	}
+
+	return &port, nil
+}