+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ### Added
 
+- [PR #291](https://github.com/konpyutaika/nifikop/pull/291) - **[Plugin]** Implementation on NiFiKop's plugin.
+- [PR #291](https://github.com/konpyutaika/nifikop/pull/291) - **[Operator/NifiConnection]** Implementation on NifiConnection controller.
+- 
 ### Changed
 
-- [PR #290](https://github.com/konpyutaika/nifikop/pull/290) - **[Operator/NifiCluster]** Change default sensitive algorithm
+- [PR #290](https://github.com/konpyutaika/nifikop/pull/290) - **[Operator/NifiCluster]** Change default sensitive algorithm.
 
 ### Fixed Bugs